extract board_utils from ConditionExampleGenerator
